add: Add Admin front for Product, Category and Types Product Models

+++ backend/src/models/OrderProductsModel.js
@@ -22,7 +22,7 @@ OrderProductsModel.init(
       }
     },
     qte: {
-      type: DataTypes.DECIMAL(5, 2),
+      type: DataTypes.SMALLINT,
       allowNull: false
     },
     price: {
